# indexer: configure PG work mem for perf